backend/requirements.txt
astroid==2.0.4
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.5.2
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud-language==1.1.0
googleapis-common-protos==1.5.5
grpcio==1.16.1
httplib2==0.12.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
meetup-api==0.1.1
oauthlib==2.1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
